+++ src/main/java/dev/ftb/mods/ftbevolutioncompanion/content/BeastItem.java
@@ -1,0 +1,26 @@
+package dev.ftb.mods.ftbevolutioncompanion.content;
+
+import net.minecraft.ChatFormatting;
+import net.minecraft.network.chat.Component;
+import net.minecraft.world.item.Item;
+import net.minecraft.world.item.ItemStack;
+import net.minecraft.world.item.TooltipFlag;
+import net.minecraft.world.item.component.TooltipDisplay;
+
+import java.util.function.Consumer;
+
+public class BeastItem extends Item {
+    private final String tooltipKey;
+
+    public BeastItem(Properties properties, String tooltipKey) {
+        super(properties);
+        this.tooltipKey = tooltipKey;
+    }
+
+    @Override
+    public void appendHoverText(ItemStack stack, Item.TooltipContext context, TooltipDisplay display, Consumer<Component> tooltip, TooltipFlag flag) {
+        super.appendHoverText(stack, context, display, tooltip, flag);
+
+        tooltip.accept(Component.translatable(tooltipKey).withStyle(ChatFormatting.GRAY));
+    }
+}