+++ src/main/java/dev/ftb/mods/ftbevolutioncompanion/content/BeastTrophyBlockEntity.java
@@ -1,0 +1,32 @@
+package dev.ftb.mods.ftbevolutioncompanion.content;
+
+import com.geckolib.animatable.GeoBlockEntity;
+import com.geckolib.animatable.instance.AnimatableInstanceCache;
+import com.geckolib.animatable.manager.AnimatableManager;
+import com.geckolib.animation.AnimationController;
+import com.geckolib.animation.RawAnimation;
+import com.geckolib.util.GeckoLibUtil;
+
+import net.minecraft.core.BlockPos;
+import net.minecraft.world.level.block.entity.BlockEntity;
+import net.minecraft.world.level.block.state.BlockState;
+
+public class BeastTrophyBlockEntity extends BlockEntity implements GeoBlockEntity {
+    private static final RawAnimation IDLE = RawAnimation.begin().thenLoop("animation.beast_trophy.idle");
+
+    private final AnimatableInstanceCache geoCache = GeckoLibUtil.createInstanceCache(this);
+
+    public BeastTrophyBlockEntity(BlockPos pos, BlockState state) {
+        super(CompanionContent.BEAST_TROPHY_BLOCK_ENTITY.get(), pos, state);
+    }
+
+    @Override
+    public void registerControllers(AnimatableManager.ControllerRegistrar controllers) {
+        controllers.add(new AnimationController<BeastTrophyBlockEntity>("main", 0, test -> test.setAndContinue(IDLE)));
+    }
+
+    @Override
+    public AnimatableInstanceCache getAnimatableInstanceCache() {
+        return geoCache;
+    }
+}