+++ src/main/java/dev/ftb/mods/ftbevolutioncompanion/content/BeastTrophyItem.java
@@ -1,0 +1,60 @@
+package dev.ftb.mods.ftbevolutioncompanion.content;
+
+import com.geckolib.animatable.GeoItem;
+import com.geckolib.animatable.client.GeoRenderProvider;
+import com.geckolib.animatable.instance.AnimatableInstanceCache;
+import com.geckolib.animatable.instance.SingletonAnimatableInstanceCache;
+import com.geckolib.animatable.manager.AnimatableManager;
+import com.geckolib.renderer.GeoItemRenderer;
+
+import dev.ftb.mods.ftbevolutioncompanion.client.BeastTrophyItemRenderer;
+
+import net.minecraft.ChatFormatting;
+import net.minecraft.network.chat.Component;
+import net.minecraft.world.item.BlockItem;
+import net.minecraft.world.item.Item;
+import net.minecraft.world.item.ItemStack;
+import net.minecraft.world.item.TooltipFlag;
+import net.minecraft.world.item.component.TooltipDisplay;
+import net.minecraft.world.level.block.Block;
+
+import java.util.function.Consumer;
+
+public class BeastTrophyItem extends BlockItem implements GeoItem {
+    private final AnimatableInstanceCache geoCache = new SingletonAnimatableInstanceCache(this);
+
+    public BeastTrophyItem(Block block, Properties properties) {
+        super(block, properties);
+    }
+
+    @Override
+    public void appendHoverText(ItemStack stack, Item.TooltipContext context, TooltipDisplay display, Consumer<Component> tooltip, TooltipFlag flag) {
+        super.appendHoverText(stack, context, display, tooltip, flag);
+
+        tooltip.accept(Component.translatable("block.ftb.beast_trophy.tooltip").withStyle(ChatFormatting.GRAY));
+    }
+
+    @Override
+    public void registerControllers(AnimatableManager.ControllerRegistrar controllers) {
+    }
+
+    @Override
+    public AnimatableInstanceCache getAnimatableInstanceCache() {
+        return geoCache;
+    }
+
+    @Override
+    public void createGeoRenderer(Consumer<GeoRenderProvider> consumer) {
+        consumer.accept(new GeoRenderProvider() {
+            private BeastTrophyItemRenderer renderer;
+
+            @Override
+            public GeoItemRenderer<?> getGeoItemRenderer() {
+                if (renderer == null) {
+                    renderer = new BeastTrophyItemRenderer();
+                }
+                return renderer;
+            }
+        });
+    }
+}